+++ packages/api/src/settings/controllers.ts
@@ -6,8 +6,8 @@ import { TypedRequestBody, TypedResponse } from '@/shared/types';
 import { EventLogTypeKey } from '@/eventlog/types';
 import { logEvent } from '@/eventlog/utils';
 import { settingValue } from '@/shared/settings';
-import { ExportTransformerMap } from '@/period/types';
-import { getCustomExportTransformer } from '@/period/utils/export';
+import { ExportTransformer } from '@/period/types';
+import { getCustomExportTransformer } from '@/period/utils/getCustomExportTransformer';
 import { SettingsModel } from './entities';
 import { settingListTransformer, settingTransformer } from './transformers';
 import { SettingDto, SettingSetInput } from './types';
@@ -96,7 +96,7 @@ export const set = async (
 
 export const customExportTransformer = async (
   req: Request,
-  res: TypedResponse<ExportTransformerMap>
+  res: TypedResponse<ExportTransformer>
 ): Promise<void> => {
   const customExportMapSetting = (await settingValue(
     'CUSTOM_EXPORT_MAP'